% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main function.R
\name{network_bootstrap}
\alias{network_bootstrap}
\title{Parametric bootstrap inference in network formation models}
\usage{
network_bootstrap(y, X, N, bootstrap_time, data, index, link = "probit")
}
\arguments{
\item{y}{outcome variabe}

\item{X}{dependent variables}

\item{N}{size of the individuals}

\item{bootstrap_time}{number of the bootstrap times}

\item{data}{data which you want to apply}

\item{index}{name of the import side and export side}

\item{link}{probit or logit model}
}
\value{
A list of fitted results is returned.
Within this outputted list, the following elements can be found:
\item{est}{estimated coefficients from the bootstrap procedure.}
\item{sd}{stand deviation form the bootstrap procedure.}
\item{est_MLE}{estimated coefficients from MLE.}
}
\description{
R package 'Netbootstrap' is dedicated to do parametric bootstrap inference in network formation models.
}
\details{
Parametric bootstrap inference in network formation models
}
